# Dynamic CPU saturation integration