--- a/lib/collection/src/shards/local_shard.rs
+++ b/lib/collection/src/shards/local_shard.rs
@@ -136,7 +136,7 @@ impl LocalShard {
             segment_holder.clone(),
             locked_wal.clone(),
             config.optimizer_config.flush_interval_sec,
-            Some(config.optimizer_config.max_optimization_threads),
+            config.optimizer_config.max_optimization_threads,
         );
 
         let (update_sender, update_receiver) =