--- a/lib/collection/src/shards/replica_set/update.rs
+++ b/lib/collection/src/shards/replica_set/update.rs
@@ -424,7 +424,7 @@ mod tests {
         memmap_threshold: None,
         indexing_threshold: Some(50_000),
         flush_interval_sec: 30,
-        max_optimization_threads: 2,
+        max_optimization_threads: Some(2),
     };
 
     async fn new_shard_replica_set(collection_dir: &TempDir) -> ShardReplicaSet {