In <https://github.com/qdrant/qdrant/pull/3364> we made the groundwork for dynamically saturating CPU usage while optimizing.

This PR integrates the feature with additionally required public API/configuration changes. Instead of having fixed optimization limits in the configuration, everything now defaults to being dynamic (automatic).

Configuration changes are mostly described in review comments below. I've also added a `cpu_budget` property to configure what number of CPUs to saturate for optimization tasks.

The idea is to merge this as part of Qdrant 1.8.

### All Submissions:

* [x] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### New Feature Submissions:

1. [x] Does your submission pass tests?
2. [x] Have you formatted your code locally using `cargo +nightly fmt --all` command prior to submission?
3. [x] Have you checked your code using `cargo clippy --all --all-features` command?

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why you'd like us to include them?
* [x] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully ran tests with your changes locally?