@timvisee 2024-01-09
We want to merge the internals of this feature first, so that we can release it as part of a patch release. We can integrate the public API changes for the next minor release to make this feature fully functional.

Please see <https://github.com/qdrant/qdrant/pull/3364> for the internals. I'll draft this PR until it is merged.