@timvisee 2024-01-29
## Indexing benchmark

I've done some benchmarking to measure the indexing performance of this PR. I used a machine with 64 CPU cores, using the default configuration. For each test I insert a million points with indexing disabled, then I enable it again and measure indexing time. _Old_ refers to the current latest `dev` branch, while _new_ refers to this PR.

I don't have fancy graphs, but I think the list below is good enough to show off the gains.

### Results
**2 CPU cores:**
- `bfb -n 100000`:
  - Old: 40s
  - New: 40s

**8 CPU cores:**
- `bfb -n 1000000`:
  - Old: 163s
  - New: 186s (14% slower)
    - This is slower because the new approach uses 7 cores rather than 8. We do this intentionally to spare a core for handling other stuff, such as ongoing searches. Users can configure this with `storage.performance.cpu_budget`. We can also consider to change this default [here](https://github.com/qdrant/qdrant/blob/c480b3ea1172789f5f57f30af86a43d14917b5b8/lib/common/common/src/defaults.rs#L6-L23).

**16 CPU cores:**
- `bfb -n 1000000`:
  - Old: 123s
  - New: 92s (34% faster)
- `bfb -n 1000000 --segments 4`:
  - Old: 126s
  - New: 113s (12% faster)
- `bfb -n 1000000 --segments 8`:
  - Old: 123s
  - New: 92s (34% faster)

**32 CPU cores:**
- `bfb -n 1000000`:
  - Old: 121s
  - New: 54s (124% faster)
- `bfb -n 1000000 --segments 4`:
  - Old: 127s
  - New: 96s (32% faster)
- `bfb -n 1000000 --segments 8`:
  - Old: 121s
  - New: 54s (124% faster)
- `bfb -n 1000000 --segments 16`:
  - Old: 113s
  - New: 53s (113% faster)

**64 CPU cores:**
- `bfb -n 1000000`:
  - Old: 122s
  - New: 47s (160% faster)
- `bfb -n 1000000 --segments 4`:
  - Old: 129s
  - New: 74s (74% faster)
- `bfb -n 1000000 --segments 8`:
  - Old: 119s
  - New: 48s (148% faster)
- `bfb -n 1000000 --segments 16`:
  - Old: 120s
  - New: 43s (179% faster)

When using 64 cores, here we see CPU saturation in `htop` without this PR:
![Screenshot from 2024-01-29 16-10-32](https://github.com/qdrant/qdrant/assets/856222/9c03ea3e-89de-4a94-8cfe-b42a5779049a)

And here it is with this PR:
![Screenshot from 2024-01-29 16-13-03](https://github.com/qdrant/qdrant/assets/856222/0c8b50b1-f9a2-445e-9a95-c87150222834)